fix issue #1 recognise more generic method definition, fix issue #2 show popup menu on mac and linux.

+++ src/org/intellij/sequencer/ShowSequenceAction.java
@@ -146,7 +146,7 @@ public class ShowSequenceAction extends AnAction {
         }
 
         public int getCallStackDepth() {
-            return ((Integer)dialogPanel.jSpinner.getValue()).intValue();
+            return (Integer) dialogPanel.jSpinner.getValue();
         }
 
         public boolean isProjectClassesOnly() {
